Added persistence and arg parser
To make it faster, persistence objects are now used to remember what has
been previously loaded in the database. To make it easier, argument
parser has been added to be able to specify wether to update the
persistence objects, delete the database and start from scratch.
It's also possible to specify the database name on with to perform the
import and the directory to import (this one is required).